requirements.txt
fuzzywuzzy
paddlepaddle
paddleocr
PyMuPDF
opencv-python
uvicorn
pyyaml
fastapi
python-multipart